+++ release/train_tests/multinode_persistence/compute_aws.yaml
