+++ release/train_tests/xgboost_lightgbm/compute_aws_1worker.yaml
@@ -13,8 +13,8 @@ head_node_type:
 worker_node_types:
     - name: worker_node
       instance_type: m5.4xlarge
-      max_workers: 10
-      min_workers: 10
+      max_workers: 1
+      min_workers: 1
       use_spot: false
 
 aws:
@@ -24,5 +24,5 @@ aws:
             DeleteOnTermination: true
             Iops: 5000
             Throughput: 1000
-            VolumeSize: 1000
+            VolumeSize: 200
             VolumeType: gp3
